requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
web3
python-dotenv
